Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -2,7 +2,7 @@
 
 void func()
 {
-    int fl;
+    float fl;
     scanf("%f",&fl);
     printf("You entered: %.2f",fl);
 }
